+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_user_service_functions.go
@@ -216,6 +216,12 @@ func (backend *KubernetesKurtosisBackend) RegisterUserService(ctx context.Contex
 	return serviceRegistration, nil
 }
 
+// Registers a user service for each given serviceId, allocating each an IP and ServiceGUID
+func (backend *KubernetesKurtosisBackend)RegisterUserServices(ctx context.Context, enclaveId enclave.EnclaveID, serviceIds map[service.ServiceID]bool, ) (map[service.ServiceID]*service.ServiceRegistration, map[service.ServiceID]error, error){
+	return nil, nil, stacktrace.NewError("REGISTER USER SERVICES METHOD IS UNIMPLEMENTED. DON'T USE IT")
+}
+
+
 func (backend *KubernetesKurtosisBackend) StartUserService(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
@@ -367,6 +373,10 @@ func (backend *KubernetesKurtosisBackend) StartUserService(
 	shouldDestroyPod = false
 	shouldUndoServiceUpdate = false
 	return objectsAndResources.service, nil
+}
+
+func (backend *KubernetesKurtosisBackend) StartUserServices(ctx context.Context, enclaveId enclave.EnclaveID, services map[service.ServiceGUID]*backend_interface.ServiceConfig) (map[service.ServiceGUID]service.Service, map[service.ServiceGUID]error, error){
+	return nil, nil, stacktrace.NewError("START USER SERVICES METHOD IS UNIMPLEMENTED. DON'T USE IT")
 }
 
 func (backend *KubernetesKurtosisBackend) GetUserServices(