+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -317,6 +317,11 @@ func (backend *MetricsReportingKurtosisBackend) RegisterUserService(ctx context.
 	return result, nil
 }
 
+// Registers a user service for each given serviceId, allocating each an IP and ServiceGUID
+func (backend *MetricsReportingKurtosisBackend) RegisterUserServices(ctx context.Context, enclaveId enclave.EnclaveID, serviceIds map[service.ServiceID]bool, ) (map[service.ServiceID]*service.ServiceRegistration, map[service.ServiceID]error, error){
+	return nil, nil, stacktrace.NewError("REGISTER USER SERVICES METHOD IS UNIMPLEMENTED. DON'T USE IT")
+}
+
 func (backend *MetricsReportingKurtosisBackend) StartUserService(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
@@ -370,6 +375,10 @@ func (backend *MetricsReportingKurtosisBackend) StartUserService(
 		)
 	}
 	return userService, nil
+}
+
+func (backend *MetricsReportingKurtosisBackend) StartUserServices(ctx context.Context, enclaveId enclave.EnclaveID, services map[service.ServiceGUID]*backend_interface.ServiceConfig) (map[service.ServiceGUID]service.Service, map[service.ServiceGUID]error, error){
+	return nil, nil, stacktrace.NewError("START USER SERVICES METHOD IS UNIMPLEMENTED. DON'T USE IT")
 }
 
 func (backend *MetricsReportingKurtosisBackend) GetUserServices(