+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -35,6 +35,28 @@ type FilesArtifactsExpansion struct {
 	ExpanderDirpathsToServiceDirpaths map[string]string
 }
 
+// Config options for the underlying container of a service
+type ServiceConfig struct {
+	containerImageName string
+
+	privatePorts map[string]*port_spec.PortSpec
+
+	publicPorts map[string]*port_spec.PortSpec //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+
+	entrypointArgs []string
+
+	cmdArgs []string
+
+	envVars map[string]string
+
+	// Leave as nil to not do any files artifact expansion
+	filesArtifactExpansion FilesArtifactsExpansion
+
+	cpuAllocationMillicpus uint64
+
+	memoryAllocationMegabytes uint64
+}
+
 // KurtosisBackend abstracts a Kurtosis backend, which will be a container engine (Docker or Kubernetes).
 // The heuristic for "do I need a method in KurtosisBackend?" here is "will I make one or more calls to
 // the underlying container engine?"
@@ -251,6 +273,17 @@ type KurtosisBackend interface {
 		serviceId service.ServiceID,
 	) (*service.ServiceRegistration, error)
 
+	// Registers a user service for each given serviceId, allocating each an IP and ServiceGUID
+	RegisterUserServices(
+		ctx context.Context,
+		enclaveId enclave.EnclaveID,
+		serviceIds map[service.ServiceID]bool,
+	) (
+		successfulUserServiceRegistrations map[service.ServiceID]*service.ServiceRegistration, // "set" of user service IDs that were successfully registered
+		erroredUserServiceIds map[service.ServiceID]error, // "set" of user service IDs that errored when attempting to register, with the error
+		resultErr error, // represents an error with the function itself, rather than the user services
+	)
+
 	// StartUserService consumes a service registration to create a user container with the given parameters
 	StartUserService(
 		ctx context.Context,
@@ -269,6 +302,17 @@ type KurtosisBackend interface {
 	) (
 		*service.Service,
 		error,
+	)
+
+	// StartUserService consumes service registrations to create auser container for each registration, given each service config
+	StartUserServices(
+		ctx context.Context,
+		enclaveId enclave.EnclaveID,
+		services map[service.ServiceGUID]*ServiceConfig,
+	) (
+		successfulServices map[service.ServiceGUID]service.Service, // "set" of user service GUIDs that were successfully started
+		unsuccessfulServices map[service.ServiceGUID]error, // "set" of user service GUIDs that errored when attempting to start, with the error
+		resultErr error, // represents an error with the function itself, rather than the user services
 	)
 
 	// Gets user services using the given filters, returning a map of matched user services identified by their GUID