+++ container-engine-lib/docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 
+### Changes
+* Add `RegisterUserServices` and `StartUserServices` to `KurtosisBackend` interface
+
 # 0.33.1
 ### Features
 * Implemented setting `cpuAllocationMillicpus` and `memoryAllocationMegabytes` in both Docker and Kubernetes